Use uint32_t for collateral index

+++ src/evo/providertx.h
@@ -21,7 +21,7 @@ public:
 public:
     int16_t nVersion{CURRENT_VERSION}; // message version
     int32_t nProtocolVersion{-1};
-    int32_t nCollateralIndex{-1};
+    uint32_t nCollateralIndex{(uint32_t) - 1};
     CService addr;
     CKeyID keyIDOperator;
     CKeyID keyIDOwner;
@@ -53,7 +53,7 @@ typedef std::shared_ptr<CProviderTXRegisterMN> CProviderTXRegisterMNPtr;
 typedef std::shared_ptr<const CProviderTXRegisterMN> CProviderTXRegisterMNCPtr;
 
 bool CheckProviderTxRegister(const CTransaction& tx, const CBlockIndex* pindex, CValidationState& state);
-bool IsProTxCollateral(const CTransaction& tx, int n);
-int GetProTxCollateralIndex(const CTransaction& tx);
+bool IsProTxCollateral(const CTransaction& tx, uint32_t n);
+uint32_t GetProTxCollateralIndex(const CTransaction& tx);
 
 #endif//DASH_PROVIDERTX_H
